student--salary

+++ Student-Management-System/src/main/java/com/easyWay/Student_Management_System/Dto/StudentInfoDto.java
@@ -33,7 +33,7 @@ public class StudentInfoDto {
     public String admissionClass;
     public String endDate;
     public int totalFee;
-    public int remainingFees;
+    public float remainingFees;
 
     public List<StudentFeeInfo> feeInfo;
     public List<ReportCardEntity> reportCardEntities;
